% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc.core-package.R
\docType{package}
\name{amc.core-package}
\alias{amc.core}
\alias{amc.core-package}
\title{amc.core: AMC Climate Analysis Library - Core}
\description{
Provides core components for using AMC data packages.
}
\author{
\strong{Maintainer}: Matthew Coad \email{matthew_coad@hotmail.com}

}
\keyword{internal}
