% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\docType{data}
\name{amc_dataset_class}
\alias{amc_dataset_class}
\title{Class that indicates an object is an amc dataset}
\format{
An object of class \code{character} of length 1.
}
\usage{
amc_dataset_class
}
\description{
Class that indicates an object is an amc dataset
}
\examples{
amc_dataset_class \%in\% class(giss_ta_zonal_yearly_dataset)
}
\keyword{datasets}
