% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset_state.R
\docType{data}
\name{amc_dataset_status}
\alias{amc_dataset_status}
\title{List of valid values for dataset status}
\format{
An object of class \code{list} of length 4.
}
\usage{
amc_dataset_status
}
\description{
List of valid values for dataset status
}
\examples{
amc_dataset_status$ready
}
\keyword{datasets}
