% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_datasource.R
\docType{data}
\name{amc_datasource_class}
\alias{amc_datasource_class}
\title{Class that indicates an object is an amc datasource}
\format{
An object of class \code{character} of length 1.
}
\usage{
amc_datasource_class
}
\description{
Class that indicates an object is an amc datasource
}
\examples{
amc_datasource_class \%in\% class(gistemp_datasource)
}
\keyword{datasets}
