% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amctest.R
\docType{data}
\name{amctest}
\alias{amctest}
\title{Datasource used to perform internal testing of amc}
\format{
An object of class \code{amc_datasource} (inherits from \code{list}) of length 4.
}
\usage{
amctest
}
\description{
Datasource used to perform internal testing of amc
}
\keyword{datasets}
