% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amctest.R
\docType{data}
\name{amctest_downloadfail}
\alias{amctest_downloadfail}
\title{Dataset that will always fail to download}
\format{
An object of class \code{amc_dataset} (inherits from \code{list}) of length 7.
}
\usage{
amctest_downloadfail
}
\description{
Dataset that will always fail to download
}
\keyword{datasets}
