% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amctest.R
\docType{data}
\name{amctest_quick}
\alias{amctest_quick}
\title{Dataset that performs a very quick download}
\format{
An object of class \code{amc_dataset} (inherits from \code{list}) of length 8.
}
\usage{
amctest_quick
}
\description{
Dataset that performs a very quick download
}
\keyword{datasets}
