% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2earth.R
\docType{data}
\name{co2earth_co2_iac_annual_historical_dataset}
\alias{co2earth_co2_iac_annual_historical_dataset}
\title{IAC Switzerland Global CO2 Yearly Dataset}
\format{
An object of class \code{amc_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
co2earth_co2_iac_annual_historical_dataset
}
\description{
IAC Switzerland Global CO2 Yearly Dataset
}
\examples{
co2earth_co2_iac_annual_historical_dataset
}
\keyword{datasets}
