% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2earth.R
\docType{data}
\name{co2earth_co2_iac_annual_historical_dataset_code}
\alias{co2earth_co2_iac_annual_historical_dataset_code}
\title{IAC Switzerland Global CO2 Yearly Dataset Code}
\format{
An object of class \code{character} of length 1.
}
\usage{
co2earth_co2_iac_annual_historical_dataset_code
}
\description{
IAC Switzerland Global CO2 Yearly Dataset Code
}
\examples{
co2earth_co2_iac_annual_historical_dataset_code
}
\keyword{datasets}
