% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2earth.R
\docType{data}
\name{co2earth_datasource}
\alias{co2earth_datasource}
\title{Earths CO2 data source}
\format{
An object of class \code{amc_datasource} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
co2earth_datasource
}
\description{
Earths CO2 data source
}
\examples{
co2earth_datasource
}
\keyword{datasets}
