% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2earth.R
\docType{data}
\name{co2earth_datasource_code}
\alias{co2earth_datasource_code}
\title{Earths CO2 website}
\format{
An object of class \code{character} of length 1.
}
\usage{
co2earth_datasource_code
}
\description{
Earths CO2 website
}
\examples{
co2earth_datasource_code
}
\keyword{datasets}
