% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giss.R
\name{download_giss_ta_zonal_yearly}
\alias{download_giss_ta_zonal_yearly}
\title{Download met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Datasets}
\usage{
download_giss_ta_zonal_yearly()
}
\description{
Download met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Datasets
}
