% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac.R
\name{download_iac_co2_yearly_dataset}
\alias{download_iac_co2_yearly_dataset}
\title{Download IAC Switzerland Global CO2 Yearly Dataset}
\usage{
download_iac_co2_yearly_dataset()
}
\value{
Dataset state
}
\description{
Download IAC Switzerland Global CO2 Yearly Dataset
}
\examples{
download_iac_co2_yearly_dataset()
}
