% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\name{download_metuk_hadcet_daily_dataset}
\alias{download_metuk_hadcet_daily_dataset}
\title{Download Hadley Centre Central England Temperature (HadCET) Daily Dataset}
\usage{
download_metuk_hadcet_daily_dataset()
}
\value{
Dataset state
}
\description{
Download Hadley Centre Central England Temperature (HadCET) Daily Dataset
}
\examples{
download_metuk_hadcet_daily_dataset()
}
