% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\name{download_metuk_hadcet_monthly}
\alias{download_metuk_hadcet_monthly}
\title{Download}
\usage{
download_metuk_hadcet_monthly()
}
\value{
Vector of downloaded files
}
\description{
Download
}
\examples{
download_metuk_hadcet_monthly()
}
