% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\name{download_metuk_hadcet_yearly_dataset}
\alias{download_metuk_hadcet_yearly_dataset}
\title{Download met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Yearly Datasets}
\usage{
download_metuk_hadcet_yearly_dataset()
}
\value{
Dataset state
}
\description{
Download met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Yearly Datasets
}
\examples{
download_metuk_hadcet_monthly_dataset()
}
