% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncei.R
\name{download_ncei_total_solar_irradiance}
\alias{download_ncei_total_solar_irradiance}
\title{Download NCEI total solar irradiance datafiles to its local repository}
\usage{
download_ncei_total_solar_irradiance()
}
\value{
Filepath of the downloaded files
}
\description{
Download NCEI total solar irradiance datafiles to its local repository
}
\examples{
download_ncei_total_solar_irradiance()
}
