% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncei.R
\name{download_ncei_tsi}
\alias{download_ncei_tsi}
\title{Download NCEI total solar irradiance datafiles to its local repository}
\usage{
download_ncei_tsi()
}
\description{
Download NCEI total solar irradiance datafiles to its local repository
}
