% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_repository.R
\name{get_amc_dataset_path}
\alias{get_amc_dataset_path}
\title{Get path to a file within an amc dataset}
\usage{
get_amc_dataset_path(dataset, filename)
}
\arguments{
\item{dataset}{Target dataset}

\item{filename}{Files filename}
}
\value{
Character vector
}
\description{
Get path to a file within an amc dataset
}
\examples{
get_amc_dataset_path(gistemp_zonal_annual_means_dataset, "ZonAnn.Ts+dSST.csv")
}
