% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_repository.R
\name{get_amc_dataset_repository_path}
\alias{get_amc_dataset_repository_path}
\title{Get Path used to store data related to a specific dataset}
\usage{
get_amc_dataset_repository_path(dataset)
}
\arguments{
\item{dataset}{Target dataset}
}
\value{
File path as a character value
}
\description{
Get Path used to store data related to a specific dataset
}
\examples{
get_amc_dataset_repository_path(test_dataset1)
}
