% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_repository.R
\name{get_amc_datasource_repository_path}
\alias{get_amc_datasource_repository_path}
\title{Get Path used to store data related to a specific datasource}
\usage{
get_amc_datasource_repository_path(datasource)
}
\arguments{
\item{datasource}{Target datasource}
}
\value{
File path
}
\description{
Get Path used to store data related to a specific datasource
}
\examples{
get_amc_datasource_repository_path(amctest)
}
