% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_repository.R
\name{get_amc_default_repository_path}
\alias{get_amc_default_repository_path}
\title{Get the default AMC Repository path.}
\usage{
get_amc_default_repository_path()
}
\value{
File path as a character value
}
\description{
This is the path used to store locally cached amc data if not specifically overridden by the
"amc.repository.path" option.
}
\examples{
get_amc_default_repository_path()
}
