% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_observation_type.R
\name{get_amc_observation_type}
\alias{get_amc_observation_type}
\title{Get observation types}
\usage{
get_amc_observation_type()
}
\value{
Vector of observation types
}
\description{
Get observation types
}
\examples{
get_amc_observation_type()
}
