% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_repository.R
\name{get_amc_repository_path}
\alias{get_amc_repository_path}
\title{Get AMC Repository path.}
\usage{
get_amc_repository_path()
}
\value{
File path as a character value
}
\description{
AMC data will downloaded/read from within this data repository.
}
\examples{
get_amc_repository_path()
}
