% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{get_read_amc_dataset_func}
\alias{get_read_amc_dataset_func}
\title{Get underlying read dataset implementation}
\usage{
get_read_amc_dataset_func(dataset, required = FALSE)
}
\arguments{
\item{dataset}{Target Dataset}

\item{required}{If true function is required and will fail if it is missing}
}
\value{
Function that takes no parameters, NULL if function cannot be found
}
\description{
Get underlying read dataset implementation
}
