% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giss.R
\docType{data}
\name{giss}
\alias{giss}
\title{Goddard Institute for Space Studies (GISS) data source}
\format{
An object of class \code{amc_datasource} (inherits from \code{list}) of length 4.
}
\usage{
giss
}
\description{
Goddard Institute for Space Studies (GISS) data source
}
\examples{
giss
}
\keyword{datasets}
