% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giss.R
\docType{data}
\name{giss_datasource}
\alias{giss_datasource}
\title{Goddard Institute for Space Studies (GISS) data source}
\format{
An object of class \code{amc_datasource} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
giss_datasource
}
\description{
Goddard Institute for Space Studies (GISS) data source
}
\examples{
giss_datasource
}
\keyword{datasets}
