% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giss.R
\docType{data}
\name{giss_ta_zonal_yearly}
\alias{giss_ta_zonal_yearly}
\title{GISS Zonal Annual Mean Temperature Anomaly}
\format{
An object of class \code{amc_dataset} (inherits from \code{list}) of length 8.
}
\usage{
giss_ta_zonal_yearly
}
\description{
GISS Zonal Annual Mean Temperature Anomaly
}
\examples{
giss_ta_zonal_yearly
}
\keyword{datasets}
