% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giss.R
\docType{data}
\name{giss_ta_zonal_yearly_dataset}
\alias{giss_ta_zonal_yearly_dataset}
\title{GISS Zonal Annual Mean Temperature Anomaly}
\format{
An object of class \code{amc_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
giss_ta_zonal_yearly_dataset
}
\description{
GISS Zonal Annual Mean Temperature Anomaly
}
\examples{
giss_ta_zonal_yearly_dataset
}
\keyword{datasets}
