% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistemp.R
\docType{data}
\name{gistemp_datasource}
\alias{gistemp_datasource}
\title{Goddard Institute for Space Studies, GISS Surface Temperature Analysis amc data source}
\format{
An object of class \code{amc_datasource} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
gistemp_datasource
}
\description{
Goddard Institute for Space Studies, GISS Surface Temperature Analysis amc data source
}
\examples{
gistemp_datasource
}
\keyword{datasets}
