% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistemp.R
\docType{data}
\name{gistemp_datasource_code}
\alias{gistemp_datasource_code}
\title{Gistemp datasource code}
\format{
An object of class \code{character} of length 1.
}
\usage{
gistemp_datasource_code
}
\description{
Gistemp datasource code
}
\examples{
gistemp_datasource_code
}
\keyword{datasets}
