% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistemp.R
\docType{data}
\name{gistemp_zonal_annual_means_dataset}
\alias{gistemp_zonal_annual_means_dataset}
\title{Zonal Annual Mean Temperature Anomaly, 1880-present}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
gistemp_zonal_annual_means_dataset
}
\description{
Zonal Annual Mean Temperature Anomaly, 1880-present
}
\examples{
gistemp_zonal_annual_means_dataset
}
\keyword{datasets}
