% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistemp.R
\docType{data}
\name{gistemp_zonal_annual_means_dataset_code}
\alias{gistemp_zonal_annual_means_dataset_code}
\title{Gistemp Zonal annual means}
\format{
An object of class \code{character} of length 1.
}
\usage{
gistemp_zonal_annual_means_dataset_code
}
\description{
Gistemp Zonal annual means
}
\examples{
gistemp_zonal_annual_means_dataset_code
}
\keyword{datasets}
