% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_zone.R
\docType{data}
\name{global_zone_code}
\alias{global_zone_code}
\title{Global zone code}
\format{
An object of class \code{character} of length 1.
}
\usage{
global_zone_code
}
\description{
Global zone code
}
\examples{
global_zone_code
}
\keyword{datasets}
