% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_zone.R
\docType{data}
\name{hemisphere_zone_type_code}
\alias{hemisphere_zone_type_code}
\title{Hemisphere zone type code}
\format{
An object of class \code{character} of length 1.
}
\usage{
hemisphere_zone_type_code
}
\description{
Hemisphere zone type code
}
\examples{
hemisphere_zone_type_code
}
\keyword{datasets}
