% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac.R
\docType{data}
\name{iac}
\alias{iac}
\title{IAC Switzerland data source}
\format{
An object of class \code{amc_datasource} (inherits from \code{list}) of length 4.
}
\usage{
iac
}
\description{
IAC Switzerland data source
}
\examples{
iac
}
\keyword{datasets}
