% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac.R
\docType{data}
\name{iac_annual_historical_co2_code}
\alias{iac_annual_historical_co2_code}
\title{IAC Switzerland Annual historical CO2 Dataset Code}
\format{
An object of class \code{character} of length 1.
}
\usage{
iac_annual_historical_co2_code
}
\description{
IAC Switzerland Annual historical CO2 Dataset Code
}
\examples{
iac_annual_historical_co2_code
}
\keyword{datasets}
