% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac.R
\docType{data}
\name{iac_annual_historical_co2_dataset}
\alias{iac_annual_historical_co2_dataset}
\title{IAC Switzerland Annual Historical Dataset}
\format{
An object of class \code{amc_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
iac_annual_historical_co2_dataset
}
\description{
IAC Switzerland Annual Historical Dataset
}
\examples{
iac_annual_historical_co2_dataset
}
\keyword{datasets}
