% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac.R
\docType{data}
\name{iac_co2_yearly}
\alias{iac_co2_yearly}
\title{IAC Switzerland Global CO2 Yearly Dataset}
\format{
An object of class \code{amc_dataset} (inherits from \code{list}) of length 8.
}
\usage{
iac_co2_yearly
}
\description{
IAC Switzerland Global CO2 Yearly Dataset
}
\examples{
iac_co2_yearly
}
\keyword{datasets}
