% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac.R
\docType{data}
\name{iac_co2_yearly_dataset}
\alias{iac_co2_yearly_dataset}
\title{IAC Switzerland Global CO2 Yearly Dataset}
\format{
An object of class \code{amc_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
iac_co2_yearly_dataset
}
\description{
IAC Switzerland Global CO2 Yearly Dataset
}
\examples{
iac_co2_yearly_dataset
}
\keyword{datasets}
