% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac.R
\docType{data}
\name{iac_datasource}
\alias{iac_datasource}
\title{IAC Switzerland data source}
\format{
An object of class \code{amc_datasource} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
iac_datasource
}
\description{
IAC Switzerland data source
}
\examples{
iac_datasource
}
\keyword{datasets}
