% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac.R
\docType{data}
\name{iac_datasource_code}
\alias{iac_datasource_code}
\title{IAC datasource code}
\format{
An object of class \code{character} of length 1.
}
\usage{
iac_datasource_code
}
\description{
IAC datasource code
}
\examples{
iac_datasource_code
}
\keyword{datasets}
