% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issi.R
\docType{data}
\name{issi}
\alias{issi}
\title{International Space Science Institute (ISSI) datasource}
\format{
An object of class \code{amc_datasource} (inherits from \code{list}) of length 4.
}
\usage{
issi
}
\description{
International Space Science Institute (ISSI) datasource
}
\examples{
issi
}
\keyword{datasets}
