% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issi.R
\docType{data}
\name{issi_datasource}
\alias{issi_datasource}
\title{International Space Science Institute (ISSI) datasource}
\format{
An object of class \code{amc_datasource} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
issi_datasource
}
\description{
International Space Science Institute (ISSI) datasource
}
\examples{
issi_datasource
}
\keyword{datasets}
