% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issi.R
\docType{data}
\name{issi_tsi_comp_dataset}
\alias{issi_tsi_comp_dataset}
\title{Total Solar Irradiance Composite data set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
issi_tsi_comp_dataset
}
\description{
Total Solar Irradiance Composite data set
}
\examples{
issi_tsi_comp_dataset
}
\keyword{datasets}
