% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issi.R
\docType{data}
\name{issi_tsi_comp_dataset_code}
\alias{issi_tsi_comp_dataset_code}
\title{Total Solar Irradiance Composite data set code}
\format{
An object of class \code{character} of length 1.
}
\usage{
issi_tsi_comp_dataset_code
}
\description{
Total Solar Irradiance Composite data set code
}
\examples{
issi_tsi_comp_dataset_code
}
\keyword{datasets}
