% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issi.R
\docType{data}
\name{issi_tsicomp}
\alias{issi_tsicomp}
\title{ISSI Total Solar Irradiance Composite data set}
\format{
An object of class \code{amc_dataset} (inherits from \code{list}) of length 8.
}
\usage{
issi_tsicomp
}
\description{
ISSI Total Solar Irradiance Composite data set
}
\examples{
issi_tsicomp
}
\keyword{datasets}
