% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issi.R
\docType{data}
\name{issi_tsicomp_dataset}
\alias{issi_tsicomp_dataset}
\title{ISSI Total Solar Irradiance Composite data set}
\format{
An object of class \code{amc_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
issi_tsicomp_dataset
}
\description{
ISSI Total Solar Irradiance Composite data set
}
\examples{
issi_tsicomp_dataset
}
\keyword{datasets}
