% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_datasource.R
\name{list_amc_datasource}
\alias{list_amc_datasource}
\title{List all available amc datasources}
\usage{
list_amc_datasource(envir = parent.frame())
}
\arguments{
\item{envir}{Where to look for datasets objects. Defaults to parent frame.}
}
\value{
Tibble
}
\description{
List all available amc datasources
}
\examples{
\donttest{
  list_amc_datasource()
}
}
