% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2earth.R
\name{list_co2earth_dataset}
\alias{list_co2earth_dataset}
\title{List all datasets available from the Co2 Earth dataset}
\usage{
list_co2earth_dataset()
}
\value{
Dataset tibble
}
\description{
List all datasets available from the Co2 Earth dataset
}
\examples{
list_co2earth_dataset()
}
