% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistemp.R
\name{list_gistemp_dataset}
\alias{list_gistemp_dataset}
\title{List all datasets available from the gistemp datasource}
\usage{
list_gistemp_dataset()
}
\value{
Dataset tibble
}
\description{
List all datasets available from the gistemp datasource
}
\examples{
list_gistemp_dataset()
}
