% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temprecord.R
\name{list_temprecord_dataset}
\alias{list_temprecord_dataset}
\title{List all datasets available from the Temperature record data set}
\usage{
list_temprecord_dataset()
}
\value{
Dataset tibble
}
\description{
List all datasets available from the Temperature record data set
}
\examples{
list_temprecord_dataset()
}
