% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_zone.R
\name{list_zone}
\alias{list_zone}
\title{list zone codes}
\usage{
list_zone()
}
\value{
Tibbler containing zone_type_code, zone_code and zone_name columns
}
\description{
list zone codes
}
\examples{
list_zone()
}
