% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_zone.R
\name{list_zone_type}
\alias{list_zone_type}
\title{list zone type codes}
\usage{
list_zone_type()
}
\value{
Tibble containing zone_type_code and zone_type_name columns
}
\description{
list zone type codes
}
\examples{
list_zone_type()
}
