% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\docType{data}
\name{metuk}
\alias{metuk}
\title{Met Office United Kingdom data source}
\format{
An object of class \code{amc_datasource} (inherits from \code{list}) of length 4.
}
\usage{
metuk
}
\description{
Met Office United Kingdom data source
}
\examples{
metuk
}
\keyword{datasets}
