% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\docType{data}
\name{metuk_datasource}
\alias{metuk_datasource}
\title{Met Office United Kingdom datasource}
\format{
An object of class \code{amc_datasource} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
metuk_datasource
}
\description{
Met Office United Kingdom datasource
}
\examples{
iac_datasource
}
\keyword{datasets}
