% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\docType{data}
\name{metuk_datasource_code}
\alias{metuk_datasource_code}
\title{Met Office United Kingdom datasource code}
\format{
An object of class \code{character} of length 1.
}
\usage{
metuk_datasource_code
}
\description{
Met Office United Kingdom datasource code
}
\examples{
metuk_datasource_code
}
\keyword{datasets}
