% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\docType{data}
\name{metuk_hadcet_monthly}
\alias{metuk_hadcet_monthly}
\title{Met Office Hadley Centre Central England Temperature, Mean HadCET Data, Monthly Mean data set}
\format{
An object of class \code{amc_dataset} (inherits from \code{list}) of length 8.
}
\usage{
metuk_hadcet_monthly
}
\description{
Met Office Hadley Centre Central England Temperature, Mean HadCET Data, Monthly Mean data set
}
\examples{
metuk_hadcet_monthly
}
\keyword{datasets}
