% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\docType{data}
\name{metuk_hadcet_monthly_dataset}
\alias{metuk_hadcet_monthly_dataset}
\title{Met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Monthly Dataset}
\format{
An object of class \code{amc_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
metuk_hadcet_monthly_dataset
}
\description{
Met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Monthly Dataset
}
\examples{
metuk_hadcet_monthly_dataset
}
\keyword{datasets}
