% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\docType{data}
\name{metuk_hadcet_monthly_dataset_code}
\alias{metuk_hadcet_monthly_dataset_code}
\title{Met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Monthly Dataset Code}
\format{
An object of class \code{character} of length 1.
}
\usage{
metuk_hadcet_monthly_dataset_code
}
\description{
Met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Monthly Dataset Code
}
\examples{
metuk_hadcet_monthly_dataset_code
}
\keyword{datasets}
