% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncei.R
\docType{data}
\name{ncei}
\alias{ncei}
\title{National Centers for Environment Information datasource}
\format{
An object of class \code{amc_datasource} (inherits from \code{list}) of length 4.
}
\usage{
ncei
}
\description{
National Centers for Environment Information datasource
}
\examples{
ncei
}
\keyword{datasets}
