% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncei.R
\docType{data}
\name{ncei_datasource}
\alias{ncei_datasource}
\title{National Centers for Environment Information datasource}
\format{
An object of class \code{amc_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
ncei_datasource
}
\description{
National Centers for Environment Information datasource
}
\examples{
ncei_datasource
}
\keyword{datasets}
