% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncei.R
\docType{data}
\name{ncei_datasource_code}
\alias{ncei_datasource_code}
\title{NCEI datasource code}
\format{
An object of class \code{character} of length 1.
}
\usage{
ncei_datasource_code
}
\description{
NCEI datasource code
}
\examples{
ncei_datasource_code
}
\keyword{datasets}
