% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncei.R
\docType{data}
\name{ncei_total_solar_irradiance_code}
\alias{ncei_total_solar_irradiance_code}
\title{NCEI total solar irradiance data set unique identity code}
\format{
An object of class \code{character} of length 1.
}
\usage{
ncei_total_solar_irradiance_code
}
\description{
NCEI total solar irradiance data set unique identity code
}
\examples{
ncei_total_solar_irradiance_code
}
\keyword{datasets}
