% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncei.R
\docType{data}
\name{ncei_total_solar_irradiance_dataset}
\alias{ncei_total_solar_irradiance_dataset}
\title{NCEI total solar irradiance meta information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
ncei_total_solar_irradiance_dataset
}
\description{
NCEI total solar irradiance meta information
}
\examples{
ncei_total_solar_irradiance_dataset
}
\keyword{datasets}
