% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncei.R
\docType{data}
\name{ncei_tsi}
\alias{ncei_tsi}
\title{NCEI total solar irradiance Dataset}
\format{
An object of class \code{amc_dataset} (inherits from \code{list}) of length 8.
}
\usage{
ncei_tsi
}
\description{
NCEI total solar irradiance Dataset
}
\examples{
ncei_tsi
}
\keyword{datasets}
