% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncei.R
\docType{data}
\name{ncei_tsi_dataset}
\alias{ncei_tsi_dataset}
\title{NCEI total solar irradiance Dataset}
\format{
An object of class \code{amc_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
ncei_tsi_dataset
}
\description{
NCEI total solar irradiance Dataset
}
\examples{
ncei_tsi_dataset
}
\keyword{datasets}
