% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{new_amc_dataset}
\alias{new_amc_dataset}
\title{Create new amc dataset}
\usage{
new_amc_dataset(datasource, dataset_code, dataset_name)
}
\arguments{
\item{datasource}{Datasource that owns the dataset}

\item{dataset_code}{Unique dataset code}

\item{dataset_name}{Data set name}
}
\value{
Dataset tibble
}
\description{
Create new amc dataset
}
\examples{
example_datasource <- new_amc_datasource(
  "example",
  "AMC Example datasource",
  "https://example.com/"
)
example_1_dataset <- new_amc_dataset(
  example_datasource,
  "example_1",
  "AMC Example Dataset 1"
)
}
