% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_datasource.R
\name{new_amc_datasource}
\alias{new_amc_datasource}
\title{Create new AMC data source}
\usage{
new_amc_datasource(name, title, reference_code = "", reference_url = "")
}
\arguments{
\item{name}{Name of the data source. Must be a unique short alpha+numeric lower case string}

\item{title}{Data source title Usually the name of the organization and reference code.}

\item{reference_code}{Optional short code used to refer to the data source. Often associated publically with the owing organization.}

\item{reference_url}{URL to a reference guide to the data source. Usually the owning organizations website.}
}
\value{
AMC data source list
}
\description{
Create new AMC data source
}
\examples{
new_amc_datasource("test_datasource", "TEST - AMC test resources")
}
