% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_zone.R
\docType{data}
\name{northern_hemisphere_zone_code}
\alias{northern_hemisphere_zone_code}
\title{Northern Hemisphere zone code}
\format{
An object of class \code{character} of length 1.
}
\usage{
northern_hemisphere_zone_code
}
\description{
Northern Hemisphere zone code
}
\examples{
northern_hemisphere_zone_code
}
\keyword{datasets}
