% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{prepare_amc_dataset}
\alias{prepare_amc_dataset}
\title{Non-destructively prepares a data set for use}
\usage{
prepare_amc_dataset(dataset)
}
\arguments{
\item{dataset}{Dataset to prepare}
}
\value{
Dataset state
}
\description{
Non-destructively prepares a data set for use
}
\examples{
prepare_amc_dataset(test_dataset1)
}
