% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{read_amc_dataset}
\alias{read_amc_dataset}
\title{Read amc dataset data}
\usage{
read_amc_dataset(dataset, envir = parent.frame())
}
\arguments{
\item{dataset}{Dataset to read}

\item{envir}{Where to look for datasets objects. Defaults to parent frame.}
}
\value{
Datasets data as AMC validated tibble
}
\description{
Read amc dataset data
}
\examples{
\donttest{
  read_amc_dataset(test_dataset1)
}
}
