% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset_state.R
\name{read_amc_dataset_state}
\alias{read_amc_dataset_state}
\title{Read AMC dataset state}
\usage{
read_amc_dataset_state(dataset)
}
\arguments{
\item{dataset}{Dataset to get state of}
}
\value{
List that identifues
}
\description{
Read AMC dataset state
}
\examples{
\donttest{
  read_amc_dataset_state(giss_ta_zonal_yearly_dataset)
}

}
