% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2earth.R
\name{read_co2earth_co2_iac_annual_historical_dataset}
\alias{read_co2earth_co2_iac_annual_historical_dataset}
\title{Read Total Solar Irradiance Composite dataset}
\usage{
read_co2earth_co2_iac_annual_historical_dataset()
}
\value{
Tibble
}
\description{
Read Total Solar Irradiance Composite dataset
}
\examples{
read_co2earth_co2_iac_annual_historical_dataset()
}
