% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giss.R
\name{read_giss_ta_zonal_yearly}
\alias{read_giss_ta_zonal_yearly}
\title{Read GISS Zonal Annual Mean Temperature Anomaly}
\usage{
read_giss_ta_zonal_yearly()
}
\description{
Read GISS Zonal Annual Mean Temperature Anomaly
}
