% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistemp.R
\name{read_gistemp_zonal_annual_means}
\alias{read_gistemp_zonal_annual_means}
\title{Read gistemp zonal annual means}
\usage{
read_gistemp_zonal_annual_means()
}
\value{
Gistemp zonal annual means
}
\description{
Read gistemp zonal annual means
}
\examples{
read_gistemp_zonal_annual_means()
}
