% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac.R
\name{read_iac_annual_historical_co2}
\alias{read_iac_annual_historical_co2}
\title{Read IAC Switzerland Annual Historical Dataset}
\usage{
read_iac_annual_historical_co2()
}
\value{
Tibble
}
\description{
Read IAC Switzerland Annual Historical Dataset
}
\examples{
read_iac_annual_historical_co2()
}
