% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac.R
\name{read_iac_co2_yearly}
\alias{read_iac_co2_yearly}
\title{Read IAC Switzerland Global CO2 Yearly tibble}
\usage{
read_iac_co2_yearly()
}
\value{
AMC Tibble
}
\description{
Read IAC Switzerland Global CO2 Yearly tibble
}
\examples{
read_iac_co2_yearly()
}
