% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issi.R
\name{read_issi_tsi_comp}
\alias{read_issi_tsi_comp}
\title{Read Total Solar Irradiance Composite dataset}
\usage{
read_issi_tsi_comp()
}
\value{
Tibble
}
\description{
Read Total Solar Irradiance Composite dataset
}
\examples{
read_issi_tsi_comp()
}
