% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\name{read_metuk_hadcet_monthly}
\alias{read_metuk_hadcet_monthly}
\title{Read Met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Monthly Dataset}
\usage{
read_metuk_hadcet_monthly()
}
\value{
Tibble
}
\description{
Read Met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Monthly Dataset
}
\examples{
read_metuk_hadcet_monthly()
}
