% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\name{read_metuk_hadcet_yearly_dataset}
\alias{read_metuk_hadcet_yearly_dataset}
\title{Read Met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Yearly Dataset}
\usage{
read_metuk_hadcet_yearly_dataset()
}
\description{
Read Met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Yearly Dataset
}
