% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temprecord.R
\name{read_temprecord_nhtr}
\alias{read_temprecord_nhtr}
\title{Read 2,000-Year Northern Hemisphere Temperature Reconstruction}
\usage{
read_temprecord_nhtr()
}
\value{
Tibble
}
\description{
Read 2,000-Year Northern Hemisphere Temperature Reconstruction
}
\examples{
read_temprecord_nhtr()
}
