% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datasource.R
\name{read_test_test1}
\alias{read_test_test1}
\title{Read test dataset 1}
\usage{
read_test_test1()
}
\value{
Test dataset 1 AMC tibble
}
\description{
Read test dataset 1
}
\examples{
read_test_test1()
}
