% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{resolve_amc_dataset}
\alias{resolve_amc_dataset}
\title{Resolve value that identifies a dataset}
\usage{
resolve_amc_dataset(x)
}
\arguments{
\item{x}{Dataset identifier. Possible values include the datasets code or the dataset itself}
}
\value{
AMC dataset
}
\description{
Resolve value that identifies a dataset
}
