% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{run_download_amc_dataset}
\alias{run_download_amc_dataset}
\title{Run download of an amc dataset.}
\usage{
run_download_amc_dataset(dataset, downloader, envir = parent.frame())
}
\arguments{
\item{dataset}{The dataset to download}

\item{downloader}{The script block that performs customised download}

\item{envir}{The environment in which expr is to be evaluated.}
}
\value{
Updated dataset status
}
\description{
Run download of an amc dataset.
}
\examples{
run_download_amc_dataset(test_dataset1, {
  test_datafile1_filepath <- get_amc_dataset_path(test_dataset1, test_datafile1_filename)
  fileConn<-file(test_datafile1_filepath)
  data_lines <- c(
    "test_dimension1,test_measure1,test_measure2",
    "Dimension-1-1,11,12",
    "Dimension-2-1,21,22"
  )
  writeLines(data_lines, fileConn)
  close(fileConn)
})
}
