% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset_state.R
\name{save_amc_dataset_state}
\alias{save_amc_dataset_state}
\title{Save AMC dataset state}
\usage{
save_amc_dataset_state(
  dataset,
  status,
  prepare_date = NULL,
  fail_condition = NULL
)
}
\arguments{
\item{dataset}{The dataset we are saving the state of}

\item{status}{Dataset status, Should be one of the members of the amc_dataset_status list}

\item{prepare_date}{Optional date the dataset was prepared}

\item{fail_condition}{Condition that caused dataset preparation to fail}
}
\value{
Saved state
}
\description{
Save AMC dataset state
}
\examples{
save_amc_dataset_state(test_dataset1, amc_dataset_status$ready)
}
