% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_zone.R
\docType{data}
\name{southern_hemisphere_zone_code}
\alias{southern_hemisphere_zone_code}
\title{Southern Hemisphere zone code}
\format{
An object of class \code{character} of length 1.
}
\usage{
southern_hemisphere_zone_code
}
\description{
Southern Hemisphere zone code
}
\examples{
southern_hemisphere_zone_code
}
\keyword{datasets}
