% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempr.R
\docType{data}
\name{tempr}
\alias{tempr}
\title{Temperature Record Website Datasource}
\format{
An object of class \code{amc_datasource} (inherits from \code{list}) of length 4.
}
\usage{
tempr
}
\description{
Temperature Record Website Datasource
}
\examples{
tempr
}
\keyword{datasets}
