% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempr.R
\docType{data}
\name{tempr_datasource}
\alias{tempr_datasource}
\title{Temperature Record Website Datasource}
\format{
An object of class \code{amc_datasource} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
tempr_datasource
}
\description{
Temperature Record Website Datasource
}
\examples{
tempr_datasource
}
\keyword{datasets}
