% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempr.R
\docType{data}
\name{tempr_nhtr}
\alias{tempr_nhtr}
\title{TEMPR 2,000 Year Northern Hemisphere Temperature Anomaly Temperature Reconstruction data set}
\format{
An object of class \code{amc_dataset} (inherits from \code{list}) of length 8.
}
\usage{
tempr_nhtr
}
\description{
TEMPR 2,000 Year Northern Hemisphere Temperature Anomaly Temperature Reconstruction data set
}
\examples{
tempr_nhtr
}
\keyword{datasets}
