% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempr.R
\docType{data}
\name{tempr_nhtr_dataset}
\alias{tempr_nhtr_dataset}
\title{TEMPR 2,000 Year Northern Hemisphere Temperature Anomaly Temperature Reconstruction data set}
\format{
An object of class \code{amc_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
tempr_nhtr_dataset
}
\description{
TEMPR 2,000 Year Northern Hemisphere Temperature Anomaly Temperature Reconstruction data set
}
\examples{
tempr_nhtr_dataset
}
\keyword{datasets}
