% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temprecord.R
\docType{data}
\name{temprecord_datasource}
\alias{temprecord_datasource}
\title{Temperature Record Datasource}
\format{
An object of class \code{amc_datasource} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
temprecord_datasource
}
\description{
Temperature Record Datasource
}
\examples{
temprecord_datasource
}
\keyword{datasets}
