% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temprecord.R
\docType{data}
\name{temprecord_datasource_code}
\alias{temprecord_datasource_code}
\title{Temperature record data source code}
\format{
An object of class \code{character} of length 1.
}
\usage{
temprecord_datasource_code
}
\description{
Temperature record data source code
}
\examples{
temprecord_datasource_code
}
\keyword{datasets}
