% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temprecord.R
\docType{data}
\name{temprecord_nhtr_dataset}
\alias{temprecord_nhtr_dataset}
\title{2,000 Year Northern Hemisphere Temperature Anomaly Temperature Reconstruction data set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
temprecord_nhtr_dataset
}
\description{
2,000 Year Northern Hemisphere Temperature Anomaly Temperature Reconstruction data set
}
\examples{
temprecord_nhtr_dataset
}
\keyword{datasets}
