% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temprecord.R
\docType{data}
\name{temprecord_nhtr_dataset_code}
\alias{temprecord_nhtr_dataset_code}
\title{2,000-Year Northern Hemisphere Temperature Reconstruction data set code}
\format{
An object of class \code{character} of length 1.
}
\usage{
temprecord_nhtr_dataset_code
}
\description{
2,000-Year Northern Hemisphere Temperature Reconstruction data set code
}
\examples{
temprecord_nhtr_dataset_code
}
\keyword{datasets}
