% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datasource.R
\docType{data}
\name{test_dataset1}
\alias{test_dataset1}
\title{Dataset used to perform internal testing of AMC}
\format{
An object of class \code{amc_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
test_dataset1
}
\description{
Dataset used to perform internal testing of AMC
}
\examples{
test_datasource
}
\keyword{datasets}
