% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datasource.R
\docType{data}
\name{test_dataset1_code}
\alias{test_dataset1_code}
\title{Test dataset1 code}
\format{
An object of class \code{character} of length 1.
}
\usage{
test_dataset1_code
}
\description{
Test dataset1 code
}
\examples{
test_dataset1_code
}
\keyword{datasets}
