% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datasource.R
\docType{data}
\name{test_datasource}
\alias{test_datasource}
\title{Datasource used to perform internal testing of amc}
\format{
An object of class \code{amc_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
test_datasource
}
\description{
Datasource used to perform internal testing of amc
}
\examples{
test_datasource
}
\keyword{datasets}
