% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datasource.R
\docType{data}
\name{test_datasource_code}
\alias{test_datasource_code}
\title{Test datasource code}
\format{
An object of class \code{character} of length 1.
}
\usage{
test_datasource_code
}
\description{
Test datasource code
}
\examples{
test_datasource_code
}
\keyword{datasets}
