% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{view_amc_dataset}
\alias{view_amc_dataset}
\title{View available AMC data sets as a dataframe}
\usage{
view_amc_dataset()
}
\value{
Data frame containing data sets name, title, synopsis and reference title + url
}
\description{
View available AMC data sets as a dataframe
}
\examples{
\donttest{
  view_amc_datasource()
}
}
