% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_datasource.R
\name{view_amc_datasource}
\alias{view_amc_datasource}
\title{View available AMC data sources as a dataframe}
\usage{
view_amc_datasource()
}
\value{
Data frame containing data sources name, synopsis and reference title + url
}
\description{
View available AMC data sources as a dataframe
}
\examples{
\donttest{
  view_amc_datasource()
}
}
