% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metuk.R
\name{download_metuk_hadcet_monthly}
\alias{download_metuk_hadcet_monthly}
\title{Download met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Datasets}
\usage{
download_metuk_hadcet_monthly()
}
\description{
Download met Office United Kingdom, Hadley Centre Central England Temperature (HadCET) Datasets
}
