% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{import_amc_dataset_file}
\alias{import_amc_dataset_file}
\title{Import datafile into an amc datasets repository}
\usage{
import_amc_dataset_file(dataset, filepath, envir = parent.frame())
}
\arguments{
\item{dataset}{Data to import file into}

\item{filepath}{Path to the file to import}

\item{envir}{Where to look for datasets objects. Defaults to parent frame.}
}
\value{
Import successfuly
}
\description{
Import datafile into an amc datasets repository
}
\examples{
\donttest{
  import_amc_dataset_file(test_dataset1, "ZonAnn.Ts+dSST.csv")
}
}
