% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{list_amc_dataset}
\alias{list_amc_dataset}
\title{List all available amc datasets}
\usage{
list_amc_dataset()
}
\value{
Tibble
}
\description{
List all available amc datasets
}
\examples{
\donttest{
  list_amc_dataset()
}
}
