% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{new_amc_dataset}
\alias{new_amc_dataset}
\title{Create new AMC dataset}
\usage{
new_amc_dataset(
  datasource,
  name,
  synopsis,
  title,
  reference_code = "",
  reference_url = ""
)
}
\arguments{
\item{datasource}{Data source that supplies the dataset.}

\item{name}{Unique name of the data set. Must be a unique short alpha+numeric lower case string prefixed with data set source code
separated with underscore.}

\item{synopsis}{Brief synopsis that describes the data set, including observations it contains, zone, observation frequency and time period.}

\item{title}{Title of the data set. Should match name referred to in reference documentation.}

\item{reference_code}{Optional code that reference documentation uses to refer to the dataset}

\item{reference_url}{Optional url that can be used to find reference documentation for the datraset}
}
\value{
AMC Data set list
}
\description{
Create new AMC dataset
}
\examples{
example_datasource <- new_amc_datasource(
  "example",
  "EXAMPLE datasource",
  "https://example.com/"
)
example_1_dataset <- new_amc_dataset(
  example_datasource,
  "example_1",
  "AMCEXAMPLE, Example Dataset 1",
  "AMC Exmple Dataset"
)
}
