% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{new_amc_dataset}
\alias{new_amc_dataset}
\title{Create new amc dataset}
\usage{
new_amc_dataset(datasource, dataset_code, dataset_name)
}
\arguments{
\item{datasource}{Datasource that owns the dataset}

\item{dataset_code}{Unique dataset code}

\item{dataset_name}{Data set name}
}
\value{
Dataset tibble
}
\description{
Create new amc dataset
}
\examples{
new_amc_dataset(test_datasource, "test_dataset1", "Test dataset1")
}
