% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_datasource.R
\name{new_amc_datasource}
\alias{new_amc_datasource}
\title{Create new amc datasource}
\usage{
new_amc_datasource(datasource_code, datasource_name, datasource_reference_url)
}
\arguments{
\item{datasource_code}{Code that uniquely identifies the datasource}

\item{datasource_name}{Name of the datasource}

\item{datasource_reference_url}{Url to a reference or guide to the datasource}
}
\value{
Tibble
}
\description{
Create new amc datasource
}
\examples{
new_amc_datasource(gistemp_datasource_code, "GISTEMP v4", "https://data.giss.nasa.gov/gistemp/")
}
