% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{prepare_amc_dataset}
\alias{prepare_amc_dataset}
\title{Prepares a data set for use, including creating directories and downloading the data from
its source if needed.}
\usage{
prepare_amc_dataset(dataset)
}
\arguments{
\item{dataset}{Dataset to prepare}
}
\value{
Dataset state
}
\description{
By default prepare is non-destructive and will leave existing resource in place
}
\examples{
\donttest{
  prepare_amc_dataset(amctest_quick)
}
}
