% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{prepare_amc_dataset_repository}
\alias{prepare_amc_dataset_repository}
\title{Non-destructively prepares data set repository for use}
\usage{
prepare_amc_dataset_repository(dataset)
}
\arguments{
\item{dataset}{Dataset to prepare repository for}
}
\value{
Respository path
}
\description{
Non-destructively prepares data set repository for use
}
\examples{
\donttest{
 prepare_amc_dataset_repository(giss_ta_zonal_yearly_dataset)
}
}
