% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc_dataset.R
\name{read_amc_dataset}
\alias{read_amc_dataset}
\title{Read amc dataset data}
\usage{
read_amc_dataset(dataset)
}
\arguments{
\item{dataset}{Dataset to read}
}
\value{
Datasets data as AMC validated tibble
}
\description{
Read amc dataset data
}
\examples{
read_amc_dataset(test_dataset1)
}
